requirements.txt
pytest==6.2.5
fastapi==0.68.1
uvicorn==0.15.0
requests==2.26.0
PyGithub==1.55
